requirements.txt
PyYAML>=6.0
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
Jinja2>=3.0
